[TASK] Move Site class to Domain/Site package
This pr:

* Moves the class "Site" to the package "Domain/Site"
* Moves the logic of getDefaultLanguage to the repository
* Prepares further refactoring for the site handling in TYPO3 9

Related Issue: #2238
