+++ Classes/Domain/Index/Queue/QueueItemRepository.php
@@ -25,7 +25,7 @@ namespace ApacheSolrForTypo3\Solr\Domain\Index\Queue;
  ***************************************************************/
 
 use ApacheSolrForTypo3\Solr\IndexQueue\Item;
-use ApacheSolrForTypo3\Solr\Site;
+use ApacheSolrForTypo3\Solr\Domain\Site\Site;
 use ApacheSolrForTypo3\Solr\System\Logging\SolrLogManager;
 use ApacheSolrForTypo3\Solr\System\Records\AbstractRepository;
 use Doctrine\DBAL\DBALException;
