+++ Classes/Domain/Site/SiteHashService.php
@@ -25,7 +25,7 @@ namespace ApacheSolrForTypo3\Solr\Domain\Site;
  *  This copyright notice MUST APPEAR in all copies of the script!
  ***************************************************************/
 
-use ApacheSolrForTypo3\Solr\Site;
+use ApacheSolrForTypo3\Solr\Domain\Site\Site;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
