+++ Tests/Integration/Domain/Site/Fixtures/can_create_instance_with_root_site.xml
