+++ Tests/Integration/Domain/Site/Fixtures/can_get_translations_for_root_site.xml
