+++ Tests/Integration/Domain/Site/SiteRepositoryTest.php
@@ -25,7 +25,7 @@ namespace ApacheSolrForTypo3\Solr\Tests\Integration\Domain\Site;
  ***************************************************************/
 
 use ApacheSolrForTypo3\Solr\Domain\Site\SiteRepository;
-use ApacheSolrForTypo3\Solr\Site;
+use ApacheSolrForTypo3\Solr\Domain\Site\Site;
 use ApacheSolrForTypo3\Solr\Tests\Integration\IntegrationTest;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
